+++ CHANGES.md
@@ -1,5 +1,7 @@
 [unreleased]
 * add data to `gu_api_repo_type_data`
+* add filter `gu_install_remote_install` for remote install
+* add filter `gu_api_url_type` for API URL data
 
 #### 0.3.1 / 2021-03-08
 * update namespace